+++ src/data/codec/mod.rs
@@ -30,7 +30,7 @@ pub mod hash; // Hashing algorithms (Fnv, Fx…)
 mod integrity; // Integrity codecs (Adler32, Crc32…)
 pub mod pack; // Packed data representations
 mod radix; // Radix-based encodings (Base32, Base64, Base58…)
-mod symbol; // Symbolic codes that encode data into visual marks
+pub mod symbol; // Symbolic codes that encode data into visual marks
 
 crate::structural_mods! { // _mods, _pub_mods, _crate_internals, _hidden
     _mods {
@@ -39,7 +39,6 @@ crate::structural_mods! { // _mods, _pub_mods, _crate_internals, _hidden
             encode::_all::*,
             integrity::_all::*,
             radix::_all::*,
-            symbol::_all::*,
         };
     }
     _pub_mods {
@@ -49,6 +48,7 @@ crate::structural_mods! { // _mods, _pub_mods, _crate_internals, _hidden
             // frame::_all::*,
             hash::_all::*,
             pack::_all::*,
+            symbol::_all::*,
         };
     }
     _reexports {